refactor: Consolidated child components of Products and Carousel into parent component
